+++ linkerd/client-policy/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "linkerd-client-policy"
+version = "0.1.0"
+authors = ["Linkerd Developers <cncf-linkerd-dev@lists.cncf.io>"]
+license = "Apache-2.0"
+edition = "2021"
+publish = false
+
+[dependencies]
+ahash = "0.8"
+http = "0.2"
+linkerd-addr = { path = "../addr" }
+once_cell = "1.17"
+tower = { version = "0.4.13", features = ["ready-cache", "retry", "util"] }
