+++ linkerd/service-profiles/Cargo.toml
@@ -17,6 +17,7 @@ http = "0.2"
 http-body = "0.4"
 indexmap = "1"
 linkerd-addr = { path = "../addr" }
+linkerd-client-policy = { path = "../client-policy" }
 linkerd-distribute = { path = "../distribute" }
 linkerd-dns-name = { path = "../dns/name" }
 linkerd-error = { path = "../error" }
