+++ linkerd/service-profiles/src/lib.rs
@@ -3,17 +3,14 @@
 
 use ahash::AHashSet;
 use futures::Stream;
-use linkerd_addr::{Addr, NameAddr};
+use linkerd_addr::NameAddr;
+pub use linkerd_client_policy::{Backend, Backends, LogicalAddr, LookupAddr};
 use linkerd_error::Error;
 use linkerd_proxy_api_resolve::Metadata;
-use once_cell::sync::Lazy;
 use std::{
-    fmt,
     future::Future,
     net::SocketAddr,
     pin::Pin,
-    str::FromStr,
-    sync::Arc,
     task::{Context, Poll},
 };
 use thiserror::Error;
@@ -47,23 +44,6 @@ pub struct Profile {
     pub opaque_protocol: bool,
     pub endpoint: Option<(SocketAddr, Metadata)>,
 }
-
-/// A profile lookup target.
-#[derive(Clone, Hash, Eq, PartialEq)]
-pub struct LookupAddr(pub Addr);
-
-/// A bound logical service address
-#[derive(Clone, Hash, Eq, PartialEq)]
-pub struct LogicalAddr(pub NameAddr);
-
-#[derive(Clone, PartialEq, Eq, Hash)]
-pub struct Backend {
-    pub addr: NameAddr,
-    pub weight: u32,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq, Hash)]
-pub struct Backends(Arc<[Backend]>);
 
 #[derive(Clone, Debug)]
 pub struct GetProfileService<P>(P);
@@ -192,130 +172,6 @@ impl linkerd_stack::Param<http::RouteSet> for Profile {
 impl linkerd_stack::Param<tcp::RouteSet> for Profile {
     fn param(&self) -> tcp::RouteSet {
         self.tcp_routes.clone()
-    }
-}
-
-// === impl LookupAddr ===
-
-impl fmt::Display for LookupAddr {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl fmt::Debug for LookupAddr {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "LookupAddr({})", self.0)
-    }
-}
-
-impl FromStr for LookupAddr {
-    type Err = <Addr as FromStr>::Err;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        Addr::from_str(s).map(LookupAddr)
-    }
-}
-
-impl From<Addr> for LookupAddr {
-    fn from(a: Addr) -> Self {
-        Self(a)
-    }
-}
-
-impl From<LookupAddr> for Addr {
-    fn from(LookupAddr(addr): LookupAddr) -> Addr {
-        addr
-    }
-}
-
-// === impl LogicalAddr ===
-
-impl fmt::Display for LogicalAddr {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        self.0.fmt(f)
-    }
-}
-
-impl fmt::Debug for LogicalAddr {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "LogicalAddr({})", self.0)
-    }
-}
-
-impl FromStr for LogicalAddr {
-    type Err = <NameAddr as FromStr>::Err;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        NameAddr::from_str(s).map(LogicalAddr)
-    }
-}
-
-impl From<NameAddr> for LogicalAddr {
-    fn from(na: NameAddr) -> Self {
-        Self(na)
-    }
-}
-
-impl From<LogicalAddr> for NameAddr {
-    fn from(LogicalAddr(na): LogicalAddr) -> NameAddr {
-        na
-    }
-}
-
-// === impl Backend ===
-
-impl fmt::Debug for Backend {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.debug_struct("Backend")
-            .field("addr", &format_args!("{}", self.addr))
-            .field("weight", &self.weight)
-            .finish()
-    }
-}
-
-// === impl Backends ===
-
-impl Backends {
-    #[inline]
-    pub fn iter(&self) -> std::slice::Iter<'_, Backend> {
-        self.0.iter()
-    }
-
-    #[inline]
-    pub fn is_empty(&self) -> bool {
-        self.0.is_empty()
-    }
-}
-
-impl Default for Backends {
-    fn default() -> Self {
-        static NO_BACKENDS: Lazy<Backends> = Lazy::new(|| Backends(Arc::new([])));
-        NO_BACKENDS.clone()
-    }
-}
-
-impl FromIterator<Backend> for Backends {
-    fn from_iter<I: IntoIterator<Item = Backend>>(iter: I) -> Self {
-        let targets = iter.into_iter().collect::<Vec<_>>().into();
-        Self(targets)
-    }
-}
-
-impl AsRef<[Backend]> for Backends {
-    #[inline]
-    fn as_ref(&self) -> &[Backend] {
-        &self.0
-    }
-}
-
-impl<'a> IntoIterator for &'a Backends {
-    type Item = &'a Backend;
-    type IntoIter = std::slice::Iter<'a, Backend>;
-
-    #[inline]
-    fn into_iter(self) -> Self::IntoIter {
-        self.iter()
     }
 }
